1. functions are divided into files for the convenience of writing code (rename_files, delete_files, choose_action) 2. removed unnecessary comments 3. added 2 functions (main, main_conf)  to the file main.py
